biasAdjust(diffGeneAnalysis)	R Documentation

baisAdjust performs bias adjustment of normalized data

Description

biasAdjust takes a normalized dataset and applies a multiplicative scalar derived from the data to help account for expression biases. These expression biases can come from many sources including dye bias,hybridization efficacy,changes in personnel,etc.After bias adjustment the data is ready for differential gene analysis.

Usage


biasAdjust(normalized, numSlides)

Arguments

normalized 	 normalized is a matrix of normalized microarray data.The first row consists of headers ans the first column consists of gene names.

numSlides 	 numSlides is the total number of chips of the microarray dataset.

Value

Returns a matrix of bias adjusted normalized data.

Note

biasAdjust takes 2-3 seconds to execute under optimal conditions of size of datasets and speed of the machine..Tested on a Pentium IV 1.6Ghz, 256Mb RAM with 22464 genes and 10 chips it took 12 seconds to execute.

Author(s)

Choudary L Jagarlamudi

References

Dozmorov I,Centola,M. An associative analysis of gene expression array data. Bioinformatics.2003 Jan22;19(2):204-11

Knowlton N,Dozmorov I, Centola M. Microarray data Analysis Tool box(MDAT): for normalization,adjustment and analysis of gene expression data. Bioonformatics.2004 Dec 12;20(18):3687-90

Examples


 
 #normalized is loaded in the package. Run the examples as follows:
 #Read the description file for best results.
 #data(normalized)
 #biasAdjust(normalized,7)


[Package diffGeneAnalysis version 1.10 Index]