rawdata(diffGeneAnalysis)	R Documentation

Micro array dataset

Description

Microarray dataset consisting of 7 chips.3 control and 4 experiment with 2382 genes.The first column consists of gene names, hence the dataset has 8 columns in total.

Usage

data(rawdata)

Format

A data frame with 2382 observations on the following 11 variables.

GeneID

Gene IDS or Gene names

c1

a numeric vector of gene intensities for control chip 1

c2

a numeric vector of gene intensities for control chip 2

c3

a numeric vectorof gene intensities for control chip 3

e1

a numeric vector of gene intensities for experiment chip 1

e2

a numeric vector of gene intensities for experiment chip 2

e3

a numeric vector of gene intensities for experiment chip 3

e4

a numeric vector of gene intensities for experiment chip 4

Source

Dozmorov I,Centola,M. An associative analysis of gene expression array data. Bioinformatics.2003 Jan22;19(2):204-11

Knowlton N,Dozmorov I, Centola M. Microarray data Analysis Tool box(MDAT): for normalization,adjustment and analysis of gene expression data. Bioonformatics.2004 Dec 12;20(18):3687-90

Examples


data(rawdata)

[Package diffGeneAnalysis version 1.10 Index]